apps/federal-sow-agent/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.20
pydantic-settings==2.6.1
python-docx==1.1.2
docxtpl==0.19.0
pypdf==5.1.0
pdfplumber==0.11.4
pandas==2.2.3
openpyxl==3.1.5
Pillow==11.0.0
pytesseract==0.3.13
langchain==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
itsdangerous==2.2.0
httpx==0.28.1
